feat: 비밀번호 검증 로직 위치 변경 및 테스트 픽스쳐 비밀번호 형식 적용
