+++ src/test/java/mocacong/server/service/MemberServiceTest.java
@@ -4,12 +4,15 @@ import java.util.List;
 import mocacong.server.domain.Member;
 import mocacong.server.dto.request.MemberSignUpRequest;
 import mocacong.server.exception.badrequest.DuplicateMemberException;
+import mocacong.server.exception.badrequest.InvalidPasswordException;
 import mocacong.server.exception.notfound.NotFoundMemberException;
 import mocacong.server.repository.MemberRepository;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
 import org.junit.jupiter.api.DisplayName;
 import org.junit.jupiter.api.Test;
+import org.junit.jupiter.params.ParameterizedTest;
+import org.junit.jupiter.params.provider.ValueSource;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.security.crypto.password.PasswordEncoder;
 
@@ -27,7 +30,7 @@ class MemberServiceTest {
     @DisplayName("회원을 정상적으로 가입한다")
     void signUp() {
         String expected = "kth990303@naver.com";
-        MemberSignUpRequest request = new MemberSignUpRequest(expected, "1234", "케이", "010-1234-5678");
+        MemberSignUpRequest request = new MemberSignUpRequest(expected, "a1b2c3d4", "케이", "010-1234-5678");
 
         memberService.signUp(request);
 
@@ -40,17 +43,33 @@ class MemberServiceTest {
     @DisplayName("이미 가입된 회원이 존재하면 회원 가입 시에 예외를 반환한다")
     void signUpByDuplicateMember() {
         String expected = "kth990303@naver.com";
-        MemberSignUpRequest request = new MemberSignUpRequest(expected, "1234", "케이", "010-1234-5678");
+        MemberSignUpRequest request = new MemberSignUpRequest(expected, "a1b2c3d4", "케이", "010-1234-5678");
         memberRepository.save(new Member(expected, "1234", "케이", "010-1234-5678"));
 
         assertThatThrownBy(() -> memberService.signUp(request))
                 .isInstanceOf(DuplicateMemberException.class);
     }
 
+    @ParameterizedTest
+    @DisplayName("비밀번호가 8~20자가 아니면 예외를 반환한다")
+    @ValueSource(strings = {"abcdef7", "abcdefgabcdefgabcde21"})
+    void passwordLengthValidation(String password) {
+        assertThatThrownBy(() -> memberService.signUp(new MemberSignUpRequest("kth990303@naver.com", password, "케이", "010-1234-5678")))
+                .isInstanceOf(InvalidPasswordException.class);
+    }
+
+    @ParameterizedTest
+    @DisplayName("비밀번호가 소문자, 숫자를 모두 포함하지 않으면 예외를 반환한다")
+    @ValueSource(strings = {"abcdefgh", "12345678"})
+    void passwordConfigureValidation(String password) {
+        assertThatThrownBy(() -> memberService.signUp(new MemberSignUpRequest("kth990303@naver.com", password, "케이", "010-1234-5678")))
+                .isInstanceOf(InvalidPasswordException.class);
+    }
+
     @Test
     @DisplayName("회원을 정상적으로 탈퇴한다")
     void delete() {
-        Member savedMember = memberRepository.save(new Member("kth990303@naver.com", "1234", "케이", "010-1234-5678"));
+        Member savedMember = memberRepository.save(new Member("kth990303@naver.com", "a1b2c3d4", "케이", "010-1234-5678"));
 
         memberService.delete(savedMember.getId());
